requirements.txt
# Web framework
fastapi
uvicorn[standard]

# Data processing
pandas
numpy

# Visualization
matplotlib
seaborn

# PDF generation
fpdf

# HTTP requests
requests

# Notion SDK (opcional, si usas solo requests puedes eliminar esta línea)
notion-client

# Type validation / models
pydantic

# Handling file uploads or multipart requests (opcional pero útil)
python-multipart

# Utilidad para sanitizar nombres de archivos
python-slugify